requirements.txt
torch>=1.12.0
transformers>=4.20.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
tqdm>=4.62.0
tensorboard>=2.8.0
pyyaml>=6.0.0
pytorch-lightning>=2.0.0
lifelines>=0.27.0 